Draw a triangle where the ship is
Triangle may not be oriented correctly - we'll find that out when we
start moving the ship according to its rotation when thrust is applied.
But for now, it looks good and it's roughly the right size and shape
(well, maybe we'll make it a bit smaller next time).

Also, fix the ship's rotation controls so that left and right aren't
swapped around - woops!